conditionally include pbi.mk
Link system libs dynamically on osx.

* http://lists.apple.com/archives/Unix-porting/2004/Oct/msg00015.html

    ld: library not found for -lcrt0.o
    clang: error: linker command failed with exit code 1

Solution: `-static` -> `-Bstatic`
`-Bstatic` is ignored by Darwin, so this is a good compromise.
Otherwise, we would need some platform-specific configuration.
(Update: This is no longer a problem after `-static` was dropped for
other reasons, but the conditional include is still helpful.)

Also, `-L` cannot be a non-existent directory with Darwin/Clang,
so we need to ifdef it somehow. The easiest way is to ifdef the
entire `include pbi.mk`. I would like to do the same thing for boost,
but not in this commit. Here, I am just trying to git OSX to work.